added case vault optimization
