+++ bluemira/magnets/cable.py
@@ -1,11 +1,23 @@
+# SPDX-FileCopyrightText: 2021-present M. Coleman, J. Cook, F. Franza
+# SPDX-FileCopyrightText: 2021-present I.A. Maione, S. McIntosh
+# SPDX-FileCopyrightText: 2021-present J. Morris, D. Short
+#
+# SPDX-License-Identifier: LGPL-2.1-or-later
+
+"""Cable class"""
+
+from typing import Callable
+
 import matplotlib.pyplot as plt
 import numpy as np
+from scipy.integrate import solve_ivp
+from scipy.optimize import minimize_scalar
 
-from bluemira.magnets.base import StructuralComponent, parall_r, serie_r
 from bluemira.magnets.strand import Strand
+from bluemira.magnets.utils import parall_r, serie_r
 
 
-class Cable(StructuralComponent):
+class Cable:
     def __init__(
             self,
             dx: float,
@@ -42,7 +54,7 @@ class Cable(StructuralComponent):
         name:
             cable string identifier
 
-        #todo decide if it is the case to add also the cooling material
+        #TODO: discuss if it is necessary to implement also the cooling material
         """
         self.name = name
         self._dx = dx
@@ -57,15 +69,17 @@ class Cable(StructuralComponent):
 
     @property
     def n_stab_strand(self):
+        """Number of stabilizing strands"""
         return self._n_stab_strand
 
     @n_stab_strand.setter
-    def n_stab_strand(self, value):
+    def n_stab_strand(self, value: int):
         self._n_stab_strand = int(np.ceil(value))
 
     def res(self, **kwargs):
         """
-        Cable's equivalent resistivity, computed as the parallel between strands' resistivity
+        Cable's equivalent resistivity, computed as the parallel between
+        strands' resistivity
 
         Parameters
         ----------
@@ -149,11 +163,52 @@ class Cable(StructuralComponent):
         """Total equivalent stiffness along y-axis"""
         return self.ym(**kwargs) * self.dx / self.dy
 
-    def Xx(self, **kwargs):
-        return 0
+    def optimize_n_stab_ths(
+            self,
+            t0: float,
+            tf: float,
+            initial_temperature: float,
+            target_temperature: float,
+            B: Callable,
+            I: Callable,
+            bounds: np.ndarray = None,
+            show: bool = False,
+    ):
+        """
+        Optimize the number of stabilizer strand in the superconducting cable using a
+        0-D hot spot criteria
 
-    def Yy(self, **kwargs):
-        return 0
+        Parameters
+        ----------
+            t0: float
+                initial time
+            tf: float
+                final time
+            initial_temperature: float
+                temperature [K] at initial time
+            target_temperature: float
+                target temperature [K] at final time
+            B : Callable
+                The magnetic field [T] as time function
+            I : Callable
+                The current [A] flowing through the conductor as time function
+            bounds: np.ndarray
+                lower and upper limits for the number of strand in the cable
+            show: bool
+                if True the behavior of temperature as function of time is plotted
+
+        Returns
+        -------
+            None
+
+        Notes
+        -----
+            The number of stabilizer strands in the cable is directly modified. An
+            error is raised in case the optimization process did not converge.
+        """
+        return optimize_n_stab_cable(
+            self, t0, tf, initial_temperature, target_temperature, B, I, bounds, show
+        )
 
     def plot(self, xc: float = 0, yc: float = 0, show: bool = False, ax=None, **kwargs):
         if ax is None:
@@ -242,3 +297,158 @@ class SquareCable(Cable):
 class DummySquareCable(SquareCable):
     def ym(self, **kwargs):
         return 120
+
+
+def _heat_balance_model_cable(t, T, B: Callable, I: Callable, cable: Cable):
+    """
+    Calculate the derivative of temperature (dT/dt) for the heat balance problem.
+
+    Parameters
+    ----------
+        t : float
+            The current time in seconds.
+        T : float
+            The current temperature in Celsius.
+        B : Callable
+            The magnetic field [T] as time function
+        I : Callable
+            The current [A] flowing through the conductor as time function
+        cable : Cable
+            the superconducting cable
+
+    Returns
+    -------
+        dTdt : float
+            The derivative of temperature with respect to time (dT/dt).
+    """
+    # Calculate the rate of heat generation (Joule dissipation)
+    if isinstance(T, np.ndarray):
+        T = T[0]
+
+    Q_gen = (I(t) / cable.area) ** 2 * cable.res(B=B(t), T=T)
+
+    # Calculate the rate of heat absorption by conductor components
+    Q_abs = cable.cp_v(T=T)
+
+    # Calculate the derivative of temperature with respect to time (dT/dt)
+    dTdt = Q_gen / Q_abs
+
+    return dTdt
+
+
+def _temperature_evolution(
+        t0: float,
+        tf: float,
+        initial_temperature: float,
+        B: Callable,
+        I: Callable,
+        cable: Cable,
+):
+    solution = solve_ivp(
+        _heat_balance_model_cable,
+        np.array([t0, tf]),
+        [initial_temperature],
+        args=(B, I, cable),
+        dense_output=True,
+    )
+
+    if not solution.success:
+        raise ValueError("Temperature evolution did not converged")
+
+    return solution
+
+
+def optimize_n_stab_cable(
+        cable: Cable,
+        t0: float,
+        tf: float,
+        initial_temperature: float,
+        target_temperature: float,
+        B: Callable,
+        I: Callable,
+        bounds: np.ndarray = None,
+        show: bool = False,
+):
+    """
+    Optimize the number of stabilizer strand in a superconducting cable using a 0-D hot spot criteria
+
+    Parameters
+    ----------
+        cable: Cable
+            the superconducting cable
+        t0: float
+            initial time
+        tf: float
+            final time
+        initial_temperature: float
+            temperature [K] at initial time
+        target_temperature: float
+            target temperature [K] at final time
+        B : Callable
+            The magnetic field [T] as time function
+        I : Callable
+            The current [A] flowing through the conductor as time function
+        bounds: np.ndarray
+            lower and upper limits for the number of strand in the cable
+        show: bool
+            if True the behavior of temperature as function of time is plotted
+
+    Returns
+    -------
+        None
+
+    Notes
+    -----
+        The number of stabilizer strands in the cable is directly modified. An error is raised in case the optimization
+        process did not converge.
+    """
+
+    def final_temperature_difference(
+            n_stab: int,
+            t0: float,
+            tf: float,
+            initial_temperature: float,
+            target_temperature: float,
+            B: Callable,
+            I: Callable,
+            cable: Cable,
+    ):
+        cable.n_stab_strand = n_stab
+
+        solution = _temperature_evolution(
+            t0=t0, tf=tf, initial_temperature=initial_temperature, B=B, I=I, cable=cable
+        )
+        final_T = float(solution.y[0][-1])
+        diff = abs(final_T - target_temperature)
+        return diff
+
+    method = None
+    if bounds is not None:
+        method = "bounded"
+
+    result = minimize_scalar(
+        fun=final_temperature_difference,
+        args=(t0, tf, initial_temperature, target_temperature, B, I, cable),
+        bounds=bounds,
+        method=method,
+    )
+
+    if not result.success:
+        raise ValueError(
+            "n_stab optimization did not converge. Check your input parameters or initial bracket."
+        )
+
+    solution = _temperature_evolution(t0, tf, initial_temperature, B, I, cable)
+    final_temperature = solution.y[0][-1]
+
+    print(f"Optimal n_stab: {cable.n_stab_strand}")
+    print(f"Final temperature with optimal n_stab: {final_temperature} Kelvin")
+
+    if show:
+        _, ax = plt.subplots()
+        ax.plot(solution.t, solution.y[0], "r")
+        time_steps = np.linspace(t0, tf, 100)
+        ax.plot(time_steps, solution.sol(time_steps)[0], "b")
+        plt.show()
+
+    return result
